+++ src/Mojopollo/Helpers/ArrInterface.php
@@ -1,0 +1,6 @@
+<?php namespace Mojopollo\Helpers;
+
+interface ArrInterface
+{
+
+}