+++ src/Mojopollo/Helpers/ArrServiceProvider.php
@@ -1,0 +1,38 @@
+<?php namespace Mojopollo\Helpers;
+
+use Illuminate\Support\ServiceProvider;
+
+class ArrServiceProvider extends ServiceProvider
+{
+
+  /**
+   * Indicates if loading of the provider is deferred.
+   *
+   * @var bool
+   */
+  protected $defer = true;
+
+  /**
+   * Register the service provider.
+   *
+   * @return void
+   */
+  public function register()
+  {
+    $this->app->bindShared('arr', function()
+    {
+      return new Arr;
+    });
+  }
+
+  /**
+   * Get the services provided by the provider.
+   *
+   * @return arr
+   */
+  public function provides()
+  {
+    return ['arr'];
+  }
+
+}