+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 Laravel Helpers
 =============
 
-##Change Log
+##Changelog
+
+####1.0.3
+* Adds Arr "Array" class
+
+####1.0.2
+* Adds camelCase and snakeCase to String class, udpates readme, adds tests
 
 ####1.0.1
 * File class added